requirements.txt
djangorestframework
jwt
